% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{coefs}
\alias{coefs}
\title{coefs}
\usage{
coefs(res)
}
\arguments{
\item{res}{the result object from mlr function}
}
\value{
The estimated values of coefficients
}
\description{
Print out the estimated values of coefficients for the fitted model
}
\examples{
data(longley)
obj = mlr("Employed", c("GNP.deflator","GNP","Unemployed","Armed.Forces","Population"), longley, 5)
coefs(obj)

}
