% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{mlr}
\alias{mlr}
\title{mlr}
\usage{
mlr(res, covs, df, decimal = 3)
}
\arguments{
\item{res}{the name of the response variable in the model}

\item{covs}{the names of covariates you'd like to involve in the model}

\item{df}{the name of your dataset}

\item{decimal}{number of decimals to keep, the default is 3}
}
\value{
a list of analytical results
}
\description{
Conduct basic analyses with multiple linear regression
}
\examples{
mlr("Employed", c("GNP.deflator","GNP","Unemployed","Armed.Forces","Population"), longley, 5)

}
