% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.cov.R
\name{select.cov}
\alias{select.cov}
\title{select.cov}
\usage{
select.cov(res, alpha = 0.05)
}
\arguments{
\item{res}{the result object from mlr function}

\item{alpha}{the given threshold of p-value, default is 0.05}
}
\value{
A vector of names of selected covariates
}
\description{
Select covariates passing a given threshold of significance
}
\examples{
data(longley)
obj = mlr("Employed", c("GNP.deflator","GNP","Unemployed","Armed.Forces","Population"), longley, 5)
select.cov(obj)

}
